% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hinteraction.R
\name{any_interaction}
\alias{any_interaction}
\title{Function to calculate Friedman's H-statistic}
\usage{
any_interaction(data = NULL, model, grid_size = 50, target = NULL)
}
\arguments{
\item{data}{data.frame w/e or w/o target}

\item{model}{model of class mlr}

\item{grid_size}{size of subsampled marginal variable distribution}

\item{target}{target variable, character}
}
\description{
Calculate pairwise Interaction Strengths
}
\references{
Friedman, J. H. & Popescu, B. E. (2008), 'Predictive learning via rule ensembles', Ann. Appl. Stat. 2(3), 916--954.
}
\author{
Maximilian Pichler
}
