% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModels.R
\name{buildClassification}
\alias{buildClassification}
\title{build model function}
\usage{
buildClassification(method, balance, settings)
}
\arguments{
\item{method}{method}

\item{balance}{method}

\item{settings}{settings}
}
\description{
build model function
}
