% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModels.R
\name{buildRegression}
\alias{buildRegression}
\title{build regression model function}
\usage{
buildRegression(method, settings)
}
\arguments{
\item{method}{method}

\item{settings}{settings}
}
\description{
build regression model function
}
