% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCommunity.R
\name{createCommunity}
\alias{createCommunity}
\title{createCommunity}
\usage{
createCommunity(
  a = NULL,
  b = NULL,
  z = NULL,
  community = NULL,
  response = NULL,
  positive = NULL,
  impute = TRUE,
  log = TRUE
)
}
\arguments{
\item{a, }{group a}

\item{b, }{group b}

\item{z, }{interaction matrix for a and b, rows == a and cols == b}

\item{community, }{list with communities/groups or data.frames}

\item{response, }{in case of community, provide name of response column}

\item{positive, }{in case of community, provide the positive class}

\item{impute, }{missing values in a and b will be imputed}

\item{log}{= T, log(x+1) will be applied to response (for count data)}
}
\description{
create Community object for \code{\link{runTM}}
}
