% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationNew.R
\name{createDiscrete}
\alias{createDiscrete}
\title{create discrete}
\usage{
createDiscrete(range)
}
\arguments{
\item{range}{range of discrete}

\item{nD}{number discrete}
}
\description{
create discrete
}
