% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createImpData.R
\name{createImpData}
\alias{createImpData}
\title{Impute Data}
\usage{
createImpData(X = NULL, Y = NULL)
}
\arguments{
\item{X}{Matrix with n rows species and n columns traits. First column = species}

\item{Y}{Matrix with n rows species and n columns traits. First column = species}
}
\description{
Impute Data
}
