% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCommunity.R
\name{createInter}
\alias{createInter}
\title{Create Interactions}
\usage{
createInter(impData, z, log)
}
\arguments{
\item{impData}{imputed groups or inter}

\item{z}{interaction matrix}

\item{log}{log}
}
\description{
Create Interactions
}
