% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInteractionList.R
\name{createInteractionList}
\alias{createInteractionList}
\title{Create transformd interaction list}
\usage{
createInteractionList(individualsX = NULL, individualsY = NULL,
  interactionsZ = NULL)
}
\arguments{
\item{X}{Matrix with n rows species and n columns traits. First column = species}

\item{Y}{Matrix with n rows species and n columns traits. First column = species}

\item{Z}{matrix with X*Y species. Interactions can be binaries or probabilities}
}
\description{
Create transformd interaction list
}
