% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{createInterFeatures}
\alias{createInterFeatures}
\title{Feature Eng funcion. Create interaction features, all combinations}
\usage{
createInterFeatures(df)
}
\arguments{
\item{df}{data.frame with the last variable as target}
}
\description{
Feature Eng funcion. Create interaction features, all combinations
}
