% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModel.R
\name{createModel}
\alias{createModel}
\title{Create Model}
\usage{
createModel(interactionList, method = "RF_random", parallel = T,
  tune = "random", target = NULL, iters = 5, tuning_metric = "auc")
}
\arguments{
\item{interactionList}{permuted data with binary classification for interaction}

\item{method}{Method to be used. RF, kNN, SVM or (deep)neural net. RF is default}

\item{parallel}{boolean. Parallel computing. Default = T}

\item{tune}{boolean. Model is tuned by default = T}

\item{target}{Default}
}
\description{
Create Model
}
