% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{DNNuncertanity}
\alias{DNNuncertanity}
\title{get sd for predictions}
\usage{
DNNuncertanity(result, data, inst, n = 100, logTarget = TRUE,
  correct = TRUE, negBin = FALSE)
}
\arguments{
\item{result}{previous result}

\item{data}{data for the model}

\item{inst}{resampling instance}
}
\description{
get sd for predictions
}
