% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModel.R
\name{findBest}
\alias{findBest}
\title{Find best Model of all available Models}
\usage{
findBest(interactionList, parallel = T, tune = "random",
  target = NULL, iters = 100)
}
\arguments{
\item{interactionList}{permuted data with binary classification for interaction}

\item{parallel}{boolean. Parallel computing. Default = T}

\item{tune}{boolean. Model is tuned by default = T}

\item{target}{Default}
}
\description{
Find best Model of all available Models
}
