% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_desired_balance}
\alias{get_desired_balance}
\title{Calculate optimal time factor for desired class proportion}
\usage{
get_desired_balance(inter, desire = 0.05)
}
\arguments{
\item{inter, }{Result from GetInteractions()}

\item{desire, }{desired value between 0 and 1}
}
\description{
Calculate optimal time factor for desired class proportion
}
