% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{GetInteractions}
\alias{GetInteractions}
\title{Get Interactions...}
\usage{
GetInteractions(mainTraits = 4, randomTraits = 4, inter = 5,
  plantSpeciesnumber = 100, pollinatorSpeciesnumber = 100,
  plantTraitNumber = 10, factorPlantTraitNumber = 5,
  pollinatorTraitNumber = 8, factorPollinatorTraitNumber = 5,
  abundance_dist = 1, binary = TRUE, desire = 0.05)
}
\arguments{
\item{mainTraits}{specify the number of traits that act as main effects}

\item{randomTraits}{specify the number of traits that have random effects}

\item{inter}{specify the number of interactions between traits}

\item{binary}{parameter to set the format of the resulting interaction matrix, if TRUE the values are binary, if false the nuerical probability values are given out}

\item{desire, }{desired value between 0 and 1}
}
\value{
A list containg the interaction matrix and vectors of the sampled traits
}
\description{
...
}
\examples{
# # Example files für die Hilfe
#
x <- GetInteractions()
View(x$interactionmatrix)
species <- GetSpeciesProperties()

hist(x$mainp)
hist(x$interp)
hist(x$randomp)


image(x$interactionmatrix)
heatmap(x$interactionmatrix)

image(test2)

test =  rpois(length(x$interactionmatrix), x$interactionmatrix * 1000000)
test2 = matrix(test, ncol = 100)
table(test2)
table(as.numeric(test>0))


}
\author{
Stefan Ried
}
