% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupModels.R
\name{getLearner}
\alias{getLearner}
\title{get Learner}
\usage{
getLearner(method, balanceClasses, predict.type = "prob",
  predict.treshold = NULL, extra = NULL)
}
\arguments{
\item{method}{which method}

\item{balanceClasses}{balance class type}

\item{predict.type}{predict type}

\item{extra}{settings for learner}
}
\description{
configure Learner
}
