% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupModels.R
\name{getPars}
\alias{getPars}
\title{set Parameter}
\usage{
getPars(method, settings = NULL, extra = NULL)
}
\arguments{
\item{method}{learner}

\item{settings}{modelSettings to bet set}

\item{extra}{settings}
}
\description{
set Parameter
}
