% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{GetSpeciesProperties}
\alias{GetSpeciesProperties}
\title{GetSpecies Properties function}
\usage{
GetSpeciesProperties(plantSpeciesnumber = 200,
  pollinatorSpeciesnumber = 200, plantTraitNumber = 11,
  factorPlantTraitNumber = 5, pollinatorTraitNumber = 5,
  factorPollinatorTraitNumber = 5, abundance_dist = 1)
}
\arguments{
\item{plantSpeciesnumber}{paramter to set the number of plant species}

\item{pollinatorSpeciesnumber}{paramter to set the number of pollinator species}

\item{plantTraitNumber}{parameter to set the number of plant traits on an interval scale}

\item{factorPlantTraitNumber}{parameter to set the number of ordinal plant traits}

\item{pollinatorTraitNumber}{parameter to set the number of pollinator traits on an interval scale}

\item{plant_abundance_dist}{parameter to set the plant abundance distribution. Either a numeric value, in this case the rate of the exponential distribution, or a function with parameter n for the number of samples}

\item{factorPollinatorTraitsNumber}{parameter to set the number of ordinal pollinator traits}
}
\value{
A list containing two species/trait matrices and two corresponding abundance vectors
}
\description{
...
}
\author{
Stefan Ried
}
