% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keras_conv.R, R/cnn.R
\name{makeRLearner.classif.keras_conv}
\alias{makeRLearner.classif.keras_conv}
\title{Keras_convlearner}
\usage{
makeRLearner.classif.keras_conv()

makeRLearner.classif.keras_conv()
}
\description{
Keras_convlearner

Keras_convlearner
}
