% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classCommunity.R
\name{mergeCommunities}
\alias{mergeCommunities}
\title{mergeCommunities
merge communities}
\usage{
mergeCommunities(data)
}
\arguments{
\item{data}{data to merge}
}
\description{
mergeCommunities
merge communities
}
