% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureModel.R
\name{modelSetup}
\alias{modelSetup}
\title{modelSetup
Setup function for model setup. It's possible to manually set parameters to overwrite default values and specify parameters to be tuned.}
\usage{
modelSetup(RF = NULL, SVM = NULL, dnn = NULL, RFranger = NULL,
  knn = NULL, knnFS = NULL, naive = NULL, boost = NULL,
  modelSetup = NULL, cnn = NULL, wideDeep = NULL, preDnn = NULL)
}
\arguments{
\item{RF}{rf from random Forest}

\item{SVM}{svm from}

\item{RFranger}{rf from ranger package}

\item{knn}{knn... from...}

\item{knnFS}{knn with feature selection}

\item{naive}{naive bayes algo}

\item{boost}{boost algo}

\item{keras}{keras tensorflow}
}
\description{
modelSetup
Setup function for model setup. It's possible to manually set parameters to overwrite default values and specify parameters to be tuned.
}
