% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hinteraction.R
\name{pairwise_interaction}
\alias{pairwise_interaction}
\title{Function to calculate Friedman's H-statistic}
\usage{
pairwise_interaction(data = NULL, model, grid_size = 50,
  target = NULL, pairs = list(a = c("x.1"), b = c("x.2", "x.3", "x.4",
  "x.5")))
}
\arguments{
\item{data}{data.frame w/e or w/o target}

\item{model}{model of class mlr}

\item{grid_size}{size of subsampled marginal variable distribution}

\item{target}{target variable, character name}

\item{pairs}{two fields with a and b (only vector of length(a) == 1 allowed)}
}
\description{
Calculate pairwise Interaction Strengths
}
\references{
Friedman, J. H. & Popescu, B. E. (2008), 'Predictive learning via rule ensembles', Ann. Appl. Stat. 2(3), 916--954.
}
\author{
Maximilian Pichler
}
