% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationRF.R
\name{permutationRF}
\alias{permutationRF}
\title{permutation Random Forest for interaction Analysis}
\usage{
permutationRF(model = NULL, parallel = T, data = NULL)
}
\arguments{
\item{model}{Provide Object from createTMmodel}

\item{parallel}{CPUs to use, default use all available CPU's}

\item{data}{If no Object from createTMmodel, provide X, Y and interactionlist in a List}
}
\description{
permutation Random Forest for interaction Analysis
}
