% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.PermutationRF}
\alias{plot.PermutationRF}
\title{Permutation random Forest plot}
\usage{
\method{plot}{PermutationRF}(x, best = 15, labels = T)
}
\arguments{
\item{best}{Integer, show best n Interactionseffects, default = 15}

\item{PermutationRF}{PermutationRF object from permutationRF()}
}
\description{
Permutation random Forest plot
}
