% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.TMsobol}
\alias{plot.TMsobol}
\title{Sensitivity plots}
\usage{
\method{plot}{TMsobol}(TMsobol, best = 30)
}
\arguments{
\item{best}{Integer, strongest n interactions to plot}

\item{TM}{Provide TMsobol object from sa()}
}
\description{
Sensitivity plots
}
