% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{plotOverfitting}
\alias{plotOverfitting}
\title{plotOverfitting}
\usage{
plotOverfitting(Results, methods = c("RF", "dnn", "cnn", "cforest", "boost",
  "knn", "SVM"), titles = c("RF", "dnn", "cnn", "cforest", "boost", "knn",
  "SVM"), measure = "auc", box = T)
}
\arguments{
\item{Results}{results from runTM}

\item{titles}{titles for the plots}

\item{measure}{which measurement to use}

\item{box}{boxplot or barplot}

\item{names}{of methods}
}
\description{
plotOverfitting
}
