% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotROC}
\alias{plotROC}
\title{Roc Plot}
\usage{
plotROC(TM, cutoff = F, spar = 0.7, cost_function = T, show_in_two = F)
}
\arguments{
\item{TM}{Provide TM object from createTM}

\item{cutoff}{show best cutoff point, default = F}

\item{cost_function}{show cust_function after yourden curve, default = T}

\item{smoothing}{parameter, 0 - 1, default = 1}
}
\description{
Roc Plot
}
