% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTM.R
\name{predict.TraitMatchingResult}
\alias{predict.TraitMatchingResult}
\title{Predict from a fitted TraitMatchingResult object}
\usage{
\method{predict}{TraitMatchingResult}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a model fitted by \code{\link{runTM}}}

\item{newdata}{newdata for predictions}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\description{
Predict from a fitted TraitMatchingResult object
}
