% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keras.R
\name{predictLearner.classif.keras}
\alias{predictLearner.classif.keras}
\title{PredictLearner for keras implemantation into mlr}
\usage{
predictLearner.classif.keras(.learner, .model, .newdata, split = c(20,
  20), ...)
}
\description{
PredictLearner for keras implemantation into mlr
}
