% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preDnn.R
\name{predictLearner.classif.preKeras}
\alias{predictLearner.classif.preKeras}
\title{PredictLearner for keras_seq implemantation into mlr}
\usage{
predictLearner.classif.preKeras(.learner, .model, .newdata, split = c(20,
  20), ...)
}
\description{
PredictLearner for keras_seq implemantation into mlr
}
