% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wideDeep.R
\name{predictLearner.classif.wideDeep}
\alias{predictLearner.classif.wideDeep}
\title{PredictLearner for wideDeep implemantation into mlr}
\usage{
\method{predictLearner}{classif.wideDeep}(.learner, .model, .newdata,
  split = c(20, 20), ...)
}
\description{
PredictLearner for wideDeep implemantation into mlr
}
