% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnnReg.R
\name{predictLearner.regr.keras_conv}
\alias{predictLearner.regr.keras_conv}
\title{PredictLearner for keras_conv implemantation into mlr}
\usage{
predictLearner.regr.keras_conv(.learner, .model, .newdata, ...)
}
\description{
PredictLearner for keras_conv implemantation into mlr
}
