% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnReg.R
\name{predictLearner.regr.keras_seq}
\alias{predictLearner.regr.keras_seq}
\title{PredictLearner for keras_seq implemantation into mlr}
\usage{
\method{predictLearner}{regr.keras_seq}(.learner, .model, .newdata, ...)
}
\description{
PredictLearner for keras_seq implemantation into mlr
}
