% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom.R
\name{predictLearner.regr.multiNomDnn}
\alias{predictLearner.regr.multiNomDnn}
\title{PredictLearner for keras_seq implemantation into mlr}
\usage{
\method{predictLearner}{regr.multiNomDnn}(.learner, .model, .newdata, ...)
}
\description{
PredictLearner for keras_seq implemantation into mlr
}
