% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negBin.R
\name{predictLearner.regr.negBinDnn}
\alias{predictLearner.regr.negBinDnn}
\title{PredictLearner for keras_seq implemantation into mlr}
\usage{
\method{predictLearner}{regr.negBinDnn}(.learner, .model, .newdata, ...)
}
\description{
PredictLearner for keras_seq implemantation into mlr
}
