% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{predictLearner.regr.poisson}
\alias{predictLearner.regr.poisson}
\title{PredictLearner for keras_seq implemantation into mlr}
\usage{
\method{predictLearner}{regr.poisson}(.learner, .model, .newdata, ...)
}
\description{
PredictLearner for keras_seq implemantation into mlr
}
