% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preDnnReg.R
\name{predictLearner.regr.preKeras}
\alias{predictLearner.regr.preKeras}
\title{PredictLearner for keras_seq implemantation into mlr}
\usage{
\method{predictLearner}{regr.preKeras}(.learner, .model, .newdata,
  split = c(20, 20), ...)
}
\description{
PredictLearner for keras_seq implemantation into mlr
}
