% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{print.PermutationRF}
\alias{print.PermutationRF}
\title{Print Function for PermutationRF}
\usage{
\method{print}{PermutationRF}(x)
}
\arguments{
\item{x, }{PermutationRF object from permutationRF()}
}
\description{
Print Function for PermutationRF
}
