% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureModel.R
\name{print.TMinfo}
\alias{print.TMinfo}
\title{print method for showInfo object}
\usage{
\method{print}{TMinfo}(info)
}
\arguments{
\item{info}{from showInfo}
}
\description{
print method for showInfo object
}
