% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTM.R
\name{print.TraitMatchingResult}
\alias{print.TraitMatchingResult}
\title{Print a fitted TraitMatchingResult model}
\usage{
\method{print}{TraitMatchingResult}(x, ...)
}
\arguments{
\item{x}{a model fitted by \code{\link{runTM}}}

\item{...}{optional arguments for compatibility with the generic function, no function implemented}
}
\description{
Print a fitted TraitMatchingResult model
}
