% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sa}
\alias{sa}
\title{Sensitivity Analysis for analyzing effects}
\usage{
sa(TM, which = names(TM$Result), plot = T)
}
\arguments{
\item{TM}{Provide TM object from TM()}

\item{which}{methods to analyse, analyse all by default}

\item{plot}{show plot, default is True}
}
\description{
Sensitivity Analysis for analyzing effects
}
