% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureModel.R
\name{showInfo}
\alias{showInfo}
\title{Show Info and parameters of available learners}
\usage{
showInfo(learner = NULL, models = T, onlyModels = F, onlyMeasures = F,
  onlyBC = F)
}
\arguments{
\item{learner}{default, all}

\item{models}{default TRUE, show info about all available models}

\item{onlyModels}{default FALSE, return only available models}

\item{onlyMeasures}{default FALSE, return only available Measures}
}
\description{
Show Info and parameters of available learners
}
