% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationNew.R
\name{simulate_batch}
\alias{simulate_batch}
\title{Simulate batch of plant-pollinator networks}
\usage{
simulate_batch(numberTraitsA = 6, numberTraitsB = 6, size = c(4, 10),
  NumberA = 50, NumberB = 100, one_way = TRUE, seed = 42,
  cutoff = 0.5, strength_range = c(10, 10), abundances = FALSE,
  specRange = c(0.5, 1.2))
}
\arguments{
\item{numberTraitsA}{number traits a}

\item{numberTraitsB}{number traits b}

\item{size}{nested list size}

\item{NumberA}{number of individuals A}

\item{NumberB}{number of individuals B}

\item{one_way}{one way interactions or not}

\item{seed}{seed}

\item{cutoff}{only individuals with at least one interaction is used in the final data, cutoff for cropping out too small observed networks}

\item{strength_range}{interaction strength range}
}
\description{
wrapper around \code{\link{simulateInteraction}}. Only numerical traits are supported
}
