% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationNew.R
\name{simulateInteraction}
\alias{simulateInteraction}
\title{Simulate plant-pollinator networks}
\usage{
simulateInteraction(species = NULL, main = NULL,
  inter = matrix(c("A1", "B1", "A2", "B2", "A3", "B3"), ncol = 2, nrow =
  3, byrow = T), weights = list(main = 1, inter = c(1, 4, 2)),
  reSim = NULL, setSeed = 42, ...)
}
\arguments{
\item{species}{can be of createSpecies class, optional, default NULL}

\item{main}{vector of main effects}

\item{inter}{two column matrix of trait-matching effects}

\item{weights}{list of two vectors for main and inter (trait-matching) effect weights}

\item{reSim}{output of simulateInteraction}

\item{...}{arguments passed to createSpecies function, see details}
}
\description{
Create two groups with weighted trait-matching effects and/or main effects.
}
\details{
\itemize{
\item \code{species}: If you want, you can directly the output of createSpecies, two groups (a and b) with traits. Otherwise, pass arguments to createSpecies with \code{...}
\item \code{main}: main effects. Currently not fully supported!
\item \code{inter}: matrix of intended trait-matching effects. N rows for n trait-matching effects. Pass trait-matching by trait names (i.e "A1", "B1")
\item \code{...} see \code{\link{createSpecies}} for detailed information.

}
}
