% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTM.R
\name{TM}
\alias{TM}
\title{Transform the data and fit model by tuning the Parameter.}
\usage{
TM(X = NULL, Y = NULL, Z = NULL, interactionList = NULL,
  method = "RF_random", tune = "random", tuning_metric = "auc",
  parallel = T, impute = T, target = NULL, iters = NULL)
}
\arguments{
\item{X}{matrix with n row species and n cols traits. First column = species}

\item{Y}{matrix with n row species and n cols traits. First column = species}

\item{Z}{matrix with X*Y species. Interactions can be binaries or probabilities}

\item{interactionList}{(Species+traits)*(Species+traits) interaction list with Interactionvalue}

\item{method}{Method to be used. RF, RF_random, knn, knn_FS, SVM, (deep)neural net or best for selecting best of these models. RF_random is default}

\item{tune}{Default method = random}

\item{tuning_metric}{Tuning evaluation metric, default = GAT}

\item{parallel}{boolean. Parallel computing. Default = T}

\item{impute}{boolean. Data is imputed by default = T}

\item{target}{Target value}

\item{iters}{for tuning control. Random search, default = 5}
}
\description{
The function \code{\link{TM}} imputes missing values in the input predictor matrices and transforms
the data into a format suitable for the available methods.
Cross validation is used to test the models' performances. Model parameter will be tuned by default.
}
