% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wideDeep.R
\name{trainLearner.classif.wideDeep}
\alias{trainLearner.classif.wideDeep}
\title{Keras_conv train function}
\usage{
\method{trainLearner}{classif.wideDeep}(.learner, .task, .subset,
  .weights = NULL, archFunction = "negExp", nLayer = 2L, epochs = 200,
  arch = 30L, drop = 0.25, embedding = 8, ...)
}
\description{
Keras_conv train function
}
