% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnReg.R
\name{trainLearner.regr.keras_seq}
\alias{trainLearner.regr.keras_seq}
\title{Keras_seq train function}
\usage{
\method{trainLearner}{regr.keras_seq}(.learner, .task, .subset,
  .weights = NULL, decay = 0.97, alpha = 0.2, batch = 25,
  archFunction = "negExp", bias = T, lr = 0.01, activation = "relu",
  embedding = T, classwt = NULL, epochs = 200, nLayer = 10,
  loss = "mean_squared_error", opti = "sgd", arch = 30, drop = 0.2,
  seed = NULL, logTarget = T, ...)
}
\description{
Keras_seq train function
}
