% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom.R
\name{trainLearner.regr.multiNomDnn}
\alias{trainLearner.regr.multiNomDnn}
\title{Keras_seq train function}
\usage{
trainLearner.regr.multiNomDnn(.learner, .task, .subset, .weights = NULL,
  batch = 25L, archFunction = "negExp", bias = T, lr = 0.01,
  epochs = 200, nLayer = 10, opti = "rmsprop", arch = 30L,
  drop = 0.2, seed = NULL, expValue = FALSE, plotN = 1,
  correct = FALSE, ...)
}
\description{
Keras_seq train function
}
