% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negBin.R
\name{trainLearner.regr.negBinDnn}
\alias{trainLearner.regr.negBinDnn}
\title{Keras_seq train function}
\usage{
\method{trainLearner}{regr.negBinDnn}(.learner, .task, .subset,
  .weights = NULL, batch = 25L, archFunction = "negExp", bias = T,
  lr = 0.01, epochs = 200, nLayer = 10, opti = "rmsprop", arch = 30L,
  seed = NULL, expValue = F, distribution = "poisson",
  activation = "relu", early_stopping = TRUE, Link = TRUE, ...)
}
\description{
Keras_seq train function
}
