% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preDnnReg.R
\name{trainLearner.regr.preKeras}
\alias{trainLearner.regr.preKeras}
\title{Keras_seq train function}
\usage{
\method{trainLearner}{regr.preKeras}(.learner, .task, .subset,
  .weights = NULL, archFunction = "negExp", lr = 0.01,
  activation = "relu", classwt = NULL, epochs = 200, nLayer = 10,
  bias = T, loss = "poisson", opti = "adamax", split = c(20, 20),
  arch = 30, drop = 0.2, ...)
}
\description{
Keras_seq train function
}
