% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraitMatching.R
\docType{package}
\name{TraitMatching}
\alias{TraitMatching}
\alias{TraitMatching-package}
\title{TraitMatching: A package for analysing bipartite networks}
\description{
The TraitMatching packig provides..
}
\section{TraitMatching functions}{

The functions...
}

