% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{tuneAndAggregate}
\alias{tuneAndAggregate}
\title{tuneAndAggregate
tune treshold and get measuremens}
\usage{
tuneAndAggregate(combinedResults = NULL, tuneT = tss,
  measures = list(auc, f1, bac, acc, fdr, fpr, npv, ppv, tnr, tpr),
  onlyTest = T, bc = "None")
}
\arguments{
\item{tuneT}{msr to tune treshold}

\item{measures}{measures to calculate}

\item{combined}{Results}
}
\description{
tuneAndAggregate
tune treshold and get measuremens
}
