% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{tuneAndAggregateRegression}
\alias{tuneAndAggregateRegression}
\title{tuneAndAggregateRegression
tune treshold and get measuremens}
\usage{
tuneAndAggregateRegression(combinedResults = NULL,
  measures = list(rmse, spearmanrho), onlyTest = T)
}
\arguments{
\item{combinedResults}{Results}

\item{measures}{msr to tune treshold}

\item{onlyTest}{measures to calculate}
}
\description{
tuneAndAggregateRegression
tune treshold and get measuremens
}
