% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationNew.R
\name{createSpecies}
\alias{createSpecies}
\title{Create species for simulateInter}
\usage{
createSpecies(NumberA = 20, NumberB = 40, traitsA = c(5, 5),
  traitsB = c(8, 8), rangeDiscrete = 2:8, seed = 1337, abundance = 2,
  speciesClass = NULL, specialist = T, coLin = NULL, sampling = runif,
  specRange = c(1, 2))
}
\arguments{
\item{NumberA}{number of A}

\item{NumberB}{number of B}

\item{traitsA}{vector for traitsA = NumberofNUmerical, NumberofCategorical}

\item{traitsB}{vector for traitsB}
}
\description{
Create species for simulateInter
}
