% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupModels.R
\name{getLearner}
\alias{getLearner}
\title{RF from randomForest package}
\usage{
getLearner(method, balanceClasses, predict.type = "prob",
  predict.treshold = NULL, extra = NULL)
}
\arguments{
\item{extra}{settings for learner}
}
\description{
RF from randomForest package
}
