% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{minOneInter}
\alias{minOneInter}
\title{select only rows and columns with at least one interaction}
\usage{
minOneInter(inter)
}
\arguments{
\item{inter}{interaction matrix}
}
\description{
select only rows and columns with at least one interaction
}
