% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTM.R
\name{runTM}
\alias{runTM}
\title{Fit TM model}
\usage{
runTM(
  community,
  settings = NULL,
  method = "RF",
  tune = c("random", "grid"),
  metric = c("AUC", "R2", "Spear"),
  parallel = TRUE,
  iters = 20L,
  crossValidation = list(outer = list(method = "CV", iters = 10), inner = list(method =
    "CV", iters = 3)),
  balance = c(FALSE, "Over", "Under", "SMOTE"),
  block = NULL,
  seed = 42,
  keepModels = c(TRUE, FALSE)
)
}
\arguments{
\item{community}{object of class classCommunity created by \code{\link[=createCommunity]{createCommunity()}}}

\item{settings}{list of settings for the machine learning model. See details}

\item{method}{Which ML algorithm to be used. RF, knn, SVM, DNN, CNN, boost, ngBinDNN, naive, or CNN. See details. "RF" is default}

\item{tune}{How to tune ML parameters. We support only "random" at the moment.}

\item{metric}{Which predictive measurement should be used for tuning. "AUC" is default. See details.}

\item{parallel}{boolean or numeric. See details.}

\item{iters}{Number of tuning steps}

\item{crossValidation}{List for CV plan. See details.}

\item{balance}{How to balance classes. Default is oversampling "Over".}

\item{block}{...}

\item{seed}{set seed}

\item{keepModels}{Whether to return all fitted Models or not.}
}
\description{
Fit, tune and cross-validate a TM model
}
\author{
Maximilian Pichler
}
