% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTM.R
\name{runTM}
\alias{runTM}
\title{runTM function}
\usage{
runTM(classCommunity, settings = NULL, method = "RF", tune = "random",
  tuningMetric = "auc", parallel = T, iters = 20,
  crossValidation = list(outer = list(method = "CV", iters = 10), inner =
  list(method = "CV", iters = 3)), balanceClasses = "Over", fitSpecies = F,
  returnOnlySetup = F, block = NULL, seed = NULL, keepModels = FALSE,
  normalize = T)
}
\arguments{
\item{classCommunity}{from createCommunity()}

\item{settings}{settings with lists for models}

\item{method}{Method to be used. RF, RF_random, knn, knn_FS, SVM, (deep)neural net or best for selecting best of these models. RF_random is default}

\item{tune}{Default method = random}

\item{tuningMetric}{Tuning evaluation metric, default = GAT}

\item{parallel}{boolean. Parallel computing. Default = T}

\item{iters}{for tuning control. Random search, default = 5}

\item{crossValidation}{cv...}

\item{balanceClasses}{method to balance Classes, default = "Over" for Oversampling}

\item{fitSpecies}{default = FALSE, per default species (first and second column) will not be fitted along, if TRUE both will be fitted but also vector with species columns name can be given}
}
\description{
runTM function
}
